requirements.txt
# AI-Studio Dependencies

# Core
pydantic>=2.0.0
pyyaml>=6.0

# LLM Router
litellm>=1.0.0

# OpenAI (for Sora video generation)
openai>=1.0.0

# Image processing
pillow>=10.0.0

# Template rendering
jinja2>=3.1.0

# CLI (for future CLI tool)
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
python-dotenv>=1.0.0
